get slots ok, avail to test, slots-no slots to handle
